Pi_arbiter/requirements.txt

python-socketio==5.7.2
requests
PCF8574
smbus-cffi

# optional for on pc testing
# tkinter?
# GPIOEmulator